making the intialize process happen instantaneously.

+++ examples/07_scratchpad/route/initialize.php
@@ -1,8 +1,9 @@
 <?php
 
-$available_dir = dirname(__FILE__) . DIRECTORY_SEPARATOR . 'available' . DIRECTORY_SEPARATOR;
 $enabled_dir = dirname(__FILE__) . DIRECTORY_SEPARATOR . 'enabled' . DIRECTORY_SEPARATOR;
-
+$touchfile = $enabled_dir . '.init';
+if( file_exists( $touchfile ) ) return;
+$available_dir = dirname(__FILE__) . DIRECTORY_SEPARATOR . 'available' . DIRECTORY_SEPARATOR;
 foreach( glob( $enabled_dir . '*.php') as $file ) {
     unlink( $file );
 }
@@ -11,3 +12,4 @@ foreach( glob($available_dir . '*.php') as $file ){
     $destination = $enabled_dir . $filename;
     symlink( '../available/' . $filename,  $destination);
 }
+touch( $touchfile );
